+++ lib/praxis/api_definition.rb
@@ -1,41 +1,57 @@
-require 'singleton'
 require 'forwardable'
 
 module Praxis
 
   class ApiDefinition
-    include Singleton
     extend Forwardable
 
     attr_reader :traits
     attr_reader :responses
     attr_reader :infos
     attr_reader :global_info
+    attr_reader :application    
 
     attr_accessor :versioning_scheme
 
+    def self.instance
+      i = Thread.current[:praxis_instance] || $praxis_initializing_instance
+      raise "Trying to use Praxis::ApiDefinition outside the context of a Praxis::Application" unless i
+      i.api_definition
+    end
+    
     def self.define(&block)
+      
+      definition = Praxis::Application.current_instance.api_definition
       if block.arity == 0
-        self.instance.instance_eval(&block)
+        definition.instance_eval(&block)
       else
-        yield(self.instance)
+        yield(definition)
+      end
+    end
+    
+    def define(&block)
+      if block.arity == 0
+        self.instance_eval(&block)
+      else
+        yield(self)
       end
     end
 
-    def initialize
+    def initialize(application)
+      @application = application
       @responses = Hash.new
       @traits = Hash.new
       @base_path = ''
 
-      @global_info = ApiGeneralInfo.new
+      @global_info = ApiGeneralInfo.new(application: application)
 
       @infos = Hash.new do |hash, version|
-        hash[version] = ApiGeneralInfo.new(@global_info, version: version)
+        hash[version] = ApiGeneralInfo.new(@global_info, application: application, version: version)
       end
     end
 
     def response_template(name, &block)
-      @responses[name] = Praxis::ResponseTemplate.new(name, &block)
+      @responses[name] = Praxis::ResponseTemplate.new(name, application, &block)
     end
 
     def response(name)
@@ -91,25 +107,26 @@ module Praxis
       data
     end
 
-    define do |api|
-      api.response_template :ok do |media_type: , location: nil, headers: nil, description: nil |
-        status 200
-        description( description || 'Standard response for successful HTTP requests.' )
-
-        media_type media_type
-        location location
-        headers headers if headers
-      end
-
-      api.response_template :created do |media_type: nil, location: nil, headers: nil, description: nil|
-        status 201
-        description( description || 'The request has been fulfilled and resulted in a new resource being created.' )
-
-        media_type media_type if media_type
-        location location
-        headers headers if headers
-      end
-    end
+    # CANNOT DEFINE IT AT FILE LOADING TIME: THE INSTANCE FOR THE API_DEFINITION IS NOT READY YET.
+    # define do |api|
+    #   api.response_template :ok do |media_type: , location: nil, headers: nil, description: nil |
+    #     status 200
+    #     description( description || 'Standard response for successful HTTP requests.' )
+    #
+    #     media_type media_type
+    #     location location
+    #     headers headers if headers
+    #   end
+    #
+    #   api.response_template :created do |media_type: nil, location: nil, headers: nil, description: nil|
+    #     status 201
+    #     description( description || 'The request has been fulfilled and resulted in a new resource being created.' )
+    #
+    #     media_type media_type if media_type
+    #     location location
+    #     headers headers if headers
+    #   end
+    # end
 
   end