+++ lib/praxis/api_general_info.rb
@@ -3,10 +3,11 @@ module Praxis
 
     attr_reader :version
 
-    def initialize(global_info=nil, version: nil)
+    def initialize(global_info=nil, application:, version: nil)
       @data = Hash.new
       @global_info = global_info
       @version = version
+      @application = application
 
       if @global_info.nil? # this *is* the global info
         version_with [:header, :params]
@@ -54,7 +55,7 @@ module Praxis
         get(:version_with)
       else
         if @global_info.nil? # this *is* the global info
-          Application.instance.versioning_scheme = val
+          @application.versioning_scheme = val
           set(:version_with, val)
         else
           raise "Use of version_with is only allowed in the global part of " \