+++ lib/praxis/bootloader.rb
@@ -49,7 +49,7 @@ module Praxis
       after(:app) do
         Praxis::Mapper.finalize!
         Praxis::Blueprint.finalize!
-        Praxis::ResourceDefinition.finalize!
+        Praxis::ResourceDefinition.finalize!(application: self.application)
       end
 
     end
@@ -112,10 +112,8 @@ module Praxis
     end
 
     def setup!
-      # use the Stats and Notifications plugins by default
-      use Praxis::Stats
+      # use the Notifications plugin by default
       use Praxis::Notifications
-
       run
     end