+++ lib/praxis/response.rb
@@ -63,24 +63,23 @@ module Praxis
       self.class.response_name
     end
 
-    def format!
+    def format!(config:)
     end
 
-    def encode!
+    def encode!(handlers:)
       case @body
       when Hash, Array
         # response payload is structured data; transform it into an entity using the handler
         # implied by the response's media type. If no handler is registered for this
         # name, assume JSON as a default handler.
-        handlers = Praxis::Application.instance.handlers
         handler = (content_type && handlers[content_type.handler_name]) || handlers['json']
         @body = handler.generate(@body)
       end
     end
 
-    def finish
-      format!
-      encode!
+    def finish(application:)
+      format!(config: application.config)
+      encode!(handlers: application.handlers)
 
       @body = Array(@body)